feat: Refactor chat summaries data flow to be reactive and fix caching logic

+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/chat/ChatViewModel.kt
@@ -411,7 +411,7 @@ class ChatViewModel(
     }
 
     private fun onCollectDeleteChatEvent(deleteChatEvent: DeleteChatEvent?) {
-        if (deleteChatEvent == null) return
+        if (deleteChatEvent == null || deleteChatEvent.chatId != state.value.chatId) return
         onDeleteChatSuccess()
         emitEffect(ChatScreenEffect.NavigateBack)
     }
